refactor relay decode
