+++ srml/yee-balances/Cargo.toml
@@ -18,7 +18,7 @@ srml-support = { git = "https://github.com/yeeco/substrate.git", default-feature
 system = { package = "srml-system", git = "https://github.com/yeeco/substrate.git", default-features = false, branch = "yee-v1.0" }
 yee-sharding-primitives = { package = "yee-sharding-primitives", path = "../../sharding/primitives", default-features = false }
 sharding = { package = "yee-srml-sharding", path = "../yee-sharding", default-features = false }
-yee-relay = { package = "yee-srml-relay", path = "../yee-relay", default-features = false }
+#yee-relay = { package = "yee-srml-relay", path = "../yee-relay", default-features = false }
 
 [dev-dependencies]
 runtime_io = { package = "sr-io", git = "https://github.com/yeeco/substrate.git", branch = "yee-v1.0" }
