+++ srml/yee-executive/Cargo.toml
@@ -13,7 +13,7 @@ primitives = { package = "sr-primitives", git = "https://github.com/yeeco/substr
 srml-support = { package = "srml-support", git = "https://github.com/yeeco/substrate.git", branch = "yee-v1.0", default-features = false }
 system = { package = "srml-system", git = "https://github.com/yeeco/substrate.git", branch = "yee-v1.0", default-features = false }
 balances = { package = "yee-srml-balances", path = "../yee-balances", default-features = false }
-yee-relay = { package = "yee-srml-relay", path = "../yee-relay", default-features = false }
+#yee-relay = { package = "yee-srml-relay", path = "../yee-relay", default-features = false }
 sharding = { package = "yee-srml-sharding", path = "../yee-sharding", default-features = false }
 yee-sharding-primitives = { package = "yee-sharding-primitives", path = "../../sharding/primitives", default-features = false }
