+++ srml/yee-relay/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "yee-srml-relay"
-version = "0.1.0"
+version = "0.4.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2018"
